move MongoDbPlanner and associated types to new planner package
