+++ mongodb/src/test/scala/quasar/physical/mongodb/planner.scala
@@ -22,6 +22,7 @@ import quasar.fp._
 import quasar.javascript._
 import quasar.physical.mongodb.accumulator._
 import quasar.physical.mongodb.expression._
+import quasar.physical.mongodb.planner._
 import quasar.physical.mongodb.workflow._
 import quasar.qscript.SortDir
 import quasar.sql.{fixpoint => sql, _}
